% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_climate4R_functions.R
\name{copyXYCoords}
\alias{copyXYCoords}
\title{Copy coordinates info from one grid to another}
\usage{
copyXYCoords(x, y)
}
\arguments{
\item{x}{A grid object (see loadeR package) to be modified.}

\item{y}{A grid object from which to extract the coordinates info.}
}
\value{
The x grid object with the coordinates info from the y grid object.
}
\description{
Copy coordinates info from one grid to another
}
\examples{
#TBW
}
