% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_climate4R_functions.R
\name{dictionaryLookup}
\alias{dictionaryLookup}
\title{TBW}
\usage{
dictionaryLookup(dicPath, var, time)
}
\arguments{
\item{dicPath}{TBW}

\item{var}{TBW}

\item{time}{TBW}
}
\value{
TBW
}
\description{
TBW
}
\examples{
# TBW
}
