% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_functions.R
\name{.copyXYCoords}
\alias{.copyXYCoords}
\title{Copy coordinates from one grid to another}
\usage{
.copyXYCoords(x, y)
}
\arguments{
\item{x}{A grid object (see loadeR package) to be modified.}

\item{y}{A grid object from which to extract the coordinates info.}
}
\value{
The x grid object with the coordinates info from the y grid object.
}
\description{
Copy coordinates from one grid to another
}
\keyword{internal}
