% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_functions.R
\name{.elementwise.all.equal}
\alias{.elementwise.all.equal}
\title{Check equality in all elements of two vectors}
\usage{
.elementwise.all.equal(x, y)
}
\arguments{
\item{x}{Vector}

\item{y}{Vector}
}
\value{
Logical value indicating whether all elements in the two vectors are equal or not.
}
\description{
Check equality in all elements of two vectors
}
\keyword{internal}
