% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMIP5_functions.R
\name{downscaleCMIP5}
\alias{downscaleCMIP5}
\title{Downscale CMIP5 grid data}
\usage{
downscaleCMIP5(
  rcp,
  mod,
  indir,
  uerra,
  outdir,
  vars = NULL,
  cmip5_vars = NULL,
  lon_lim = c(-11, 12),
  lat_lim = c(28, 44),
  years = 1991:2100,
  dictionary = system.file("extdata", "CMIP5_dictionary.csv", package = "dsclim"),
  method = "GLM",
  family_link = stats::gaussian(link = "identity"),
  local_pars = NULL,
  global_pars = NULL,
  spatial_pars = NULL,
  global_nc_attributes = NULL
)
}
\arguments{
\item{rcp}{TBW}

\item{mod}{TBW}

\item{indir}{TBW}

\item{uerra}{TBW}

\item{outdir}{TBW}

\item{vars}{TBW}

\item{cmip5_vars}{TBW}

\item{lon_lim}{TBW}

\item{lat_lim}{TBW}

\item{years}{TBW}

\item{dictionary}{TBW}

\item{method}{TBW}

\item{family_link}{TBW}

\item{local_pars}{TBW}

\item{global_pars}{TBW}

\item{spatial_pars}{TBW}

\item{global_nc_attributes}{TBW}
}
\value{
TBW
}
\description{
Downscale CMIP5 grid data
}
\examples{
#TBW
}
