% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_climate4R_functions.R
\name{getVerticalLevelPars}
\alias{getVerticalLevelPars}
\title{Get vertical levels in a grid.}
\usage{
getVerticalLevelPars(grid, level)
}
\arguments{
\item{grid}{Grid object (see loadeR package) from which to get the levels}

\item{level}{Level to be found within the levels in the grid object}
}
\value{
List object with two vectors inside:
\itemize{
\item level: the level searched.
\item zRange: the position for the level.
}
}
\description{
Get vertical levels in a grid.
}
\examples{
# TBW
}
