% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_climate4R_functions.R
\name{grid2spatialdf}
\alias{grid2spatialdf}
\title{Convert climate4R grid object as spatial data frame.}
\usage{
grid2spatialdf(grid)
}
\arguments{
\item{grid}{A climate4R grid object (see \code{\link[loadeR:loadGridData]{loadeR::loadGridData()}}, \code{\link[=loadUerra]{loadUerra()}}, \code{\link[=loadTrace]{loadTrace()}}, or \code{\link[=loadCMIP5]{loadCMIP5()}})}
}
\value{
A dataframe with the format of a spatial data.frame: two columns with X and Y coordinates, and additional twelve columns for the value of the variable in each month of the year.
}
\description{
Convert climate4R grid object as spatial data frame.
}
\examples{
\dontrun{
uerra <- dsclim::loadUerra(
  "Data/UERRA/UERRA-HARMONIE/2m_temperature/latlon/1961-90_2m_temperature.nc",
  "tas"
)

grid2spatialdf(uerra)
}
}
