% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMIP_functions.R
\name{loadCMIP}
\alias{loadCMIP}
\title{Load CMIP5 data into climate4R grid format}
\usage{
loadCMIP(
  var_list,
  var_new_list,
  indir = "../../Data/CMIP5/",
  rcp,
  mod,
  lonLim = NULL,
  latLim = NULL,
  years = NULL,
  dictionary = system.file("extdata", "CMIP5_dictionary.csv", package = "TraCE21kaDSR")
)
}
\arguments{
\item{var_list}{TBW}

\item{var_new_list}{TBW}

\item{indir}{TBW}

\item{rcp}{TBW}

\item{mod}{TBW}

\item{lonLim}{TBW}

\item{latLim}{TBW}

\item{years}{TBW}

\item{dictionary}{TBW}
}
\value{
TBW
}
\description{
Load CMIP5 data into climate4R grid format
}
\examples{
#TBW
}
