% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{loadHistoricalTraceGrids}
\alias{loadHistoricalTraceGrids}
\title{Title}
\usage{
loadHistoricalTraceGrids(
  files,
  vars = NULL,
  lonLim = c(-25, 25),
  latLim = c(25, 50),
  start_date = "1551-01-01",
  end_date = "1990-12-31",
  years = 1961:1990,
  dictionary = system.file("extdata", "TraCE21ka_dictionary.csv", package =
    "TraCE21kaDSR"),
  selection_vars = c("tas", "tasmax", "tasmin", "hurs@992.5561", "ps", "pr", "cld",
    "wss"),
  compute_wss = TRUE
)
}
\arguments{
\item{files}{TBW}

\item{vars}{TBW}

\item{lonLim}{TBW}

\item{latLim}{TBW}

\item{start_date}{TBW}

\item{end_date}{TBW}

\item{years}{TBW}

\item{dictionary}{TBW}

\item{selection_vars}{TBW}

\item{compute_wss}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
# historical.files <- traceFileNames(
#                         "Data/TraCE21ka/", 36)  
# data <- loadHistoricalTraceGrids(historical.files,
#                                  c("tas", "tasmax", "tasmin", "hurs@992.5561",
#                                  "ps", "pr", "cld", "u@992.5561", "v@992.5561")) 
}
