% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{loadTraceGrid}
\alias{loadTraceGrid}
\title{Title}
\usage{
loadTraceGrid(
  file,
  var,
  lon_lim = c(-25, 25),
  lat_lim = c(25, 50),
  dictionary = system.file("extdata", "TraCE21ka_dictionary.csv", package = "dsclim")
)
}
\arguments{
\item{file}{TBW}

\item{var}{TBW}

\item{lon_lim}{TBW}

\item{lat_lim}{TBW}

\item{dictionary}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
# tas <- loadTraceGrid(
#            "Data/TraCE21ka/TS/trace.01.22000-20001BP.cam2.h0.TS.0000101-0200012.nc",
#            "tas", 1, lon_lim = NULL, lat_lim=NULL)
}
