% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{loadTraceGrids}
\alias{loadTraceGrids}
\title{Title}
\usage{
loadTraceGrids(
  files,
  vars = NULL,
  lonLim = c(-25, 25),
  latLim = c(25, 50),
  dictionary = system.file("extdata", "TraCE21ka_dictionary.csv", package =
    "TraCE21kaDSR"),
  selection_vars = c("tas", "tasmax", "tasmin", "hurs@992.5561", "ps", "pr", "cld",
    "wss"),
  compute_wss = TRUE
)
}
\arguments{
\item{files}{TBW}

\item{vars}{TBW}

\item{lonLim}{TBW}

\item{latLim}{TBW}

\item{dictionary}{TBW}

\item{selection_vars}{TBW}

\item{compute_wss}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
# LGM.files <- traceFileNames(
#                         "../Data/TraCE21ka/", 1)  
# data <- loadTraceGrids(LGM.files,
#                                  c("tas", "tasmax", "tasmin", "hurs@992.5561",
#                                  "ps", "pr", "cld", "u@992.5561", "v@992.5561"), 1,
#                                  selection_vars = c("tas", "tasmax", "tasmin",
#                                                  "hurs@992.5561", "ps", "pr",
#                                                  "cld", "wss"))
}
