% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UERRA_functions.R
\name{loadUerra}
\alias{loadUerra}
\title{Load UERRA data into climate4R grid format}
\usage{
loadUerra(
  file,
  var,
  lon_lim = c(-11, 12),
  lat_lim = c(28, 44),
  dictionary = system.file("extdata", "UERRA_dictionary.csv", package = "dsclim"),
  aggr_times = NULL,
  aggr_fun = mean
)
}
\arguments{
\item{file}{A string with the file name with the UERRA data to be loaded.}

\item{var}{A string with the variable name from the file to be loaded.}

\item{lon_lim}{An optional vector with two values specifying the longitudinal limits (min and max) to be loaded.}

\item{lat_lim}{An optional vector with two values specifying the latitudinal limits (min and max) to be loaded.}

\item{dictionary}{An optional string with the filename with the dictionary for the UERRA data. By default it loads an internal dictionary from the package.}

\item{aggr_times}{An optional numeric specifying the number of cells to be aggregated.}

\item{aggr_fun}{An optional function specifying the function to be used in aggregation if aggr.times was specified.}
}
\value{
A climate4R grid data (see \code{\link[loadeR:loadGridData]{loadeR::loadGridData()}} )
}
\description{
Load UERRA data into climate4R grid format
}
\examples{
\dontrun{
file <- "Data/UERRA/UERRA-HARMONIE/2m_temperature/latlon/1961-90_2m_temperature.nc"
var <- "tas"
uerra <- loadUerra(uerra.file, local.var)
}

}
