% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_climate4R_functions.R
\name{computeWindSpeed}
\alias{computeWindSpeed}
\title{Compute wind speed from their horizontal (u) and vertical (v) components.}
\usage{
computeWindSpeed(u, v)
}
\arguments{
\item{u}{A grid object (see loadeR package)}

\item{v}{TBW}
}
\value{
TBW
}
\description{
Compute wind speed from their horizontal (u) and vertical (v) components.
}
\examples{
#TBW

}
