% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden_functions.R
\name{.copyXYCoords}
\alias{.copyXYCoords}
\title{Copy coordinates from one grid to another}
\usage{
.copyXYCoords(x, y)
}
\arguments{
\item{x}{A grid object (see \link[loadeR:loadGridData]{loadeR::loadGridData} function) to be modified.}

\item{y}{A grid object from which to extract the coordinates info.}
}
\value{
The \code{x} grid object with the coordinates info from the \code{y} grid object.
}
\description{
Copy coordinates from one grid to another
}
\keyword{internal}
