% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMIP5_functions.R
\name{downscaleCMIP5}
\alias{downscaleCMIP5}
\title{Downscale CMIP5 grid data}
\usage{
downscaleCMIP5(
  uerra,
  local_var,
  rcp,
  mod,
  indir,
  outdir,
  var_list = NULL,
  var_new_list = NULL,
  lonLim = c(-11.25, 12.5),
  latLim = c(27, 45),
  dictionary = system.file("extdata", "CMIP5_dictionary.csv", package = "dsclim"),
  spatial_pars = NULL,
  method = "GLM",
  family_link = stats::gaussian(link = "identity"),
  global_nc_attributes = NULL
)
}
\arguments{
\item{uerra}{TBW}

\item{local_var}{TBW}

\item{rcp}{TBW}

\item{mod}{TBW}

\item{indir}{TBW}

\item{outdir}{TBW}

\item{var_list}{TBW}

\item{var_new_list}{TBW}

\item{lonLim}{TBW}

\item{latLim}{TBW}

\item{dictionary}{TBW}

\item{spatial_pars}{TBW}

\item{method}{TBW}

\item{family_link}{TBW}

\item{global_nc_attributes}{TBW}
}
\value{
TBW
}
\description{
Downscale CMIP5 grid data
}
\examples{
#TBW
}
