% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMIP_functions.R
\name{downscaleCMIP5}
\alias{downscaleCMIP5}
\title{Downscale CMIP5 grid data}
\usage{
downscaleCMIP5(
  uerra,
  var_list,
  var_new_list,
  rcp,
  mod,
  lonLim = cmip5.lon,
  latLim = cmip5.lat,
  dictionary = system.file("extdata", "CMIP5_dictionary.csv", package = "TraCE21kaDSR"),
  indir = "../../Data/CMIP5/",
  outdir = "../../Output/CMIP5/",
  local.var,
  spatial.pars = cmip5.spatial.pars,
  method = "GLM",
  family.link = family.link,
  global.nc.attributes = global.nc.attributes
)
}
\arguments{
\item{uerra}{TBW}

\item{var_list}{TBW}

\item{var_new_list}{TBW}

\item{rcp}{TBW}

\item{mod}{TBW}

\item{lonLim}{TBW}

\item{latLim}{TBW}

\item{dictionary}{TBW}

\item{indir}{TBW}

\item{outdir}{TBW}

\item{local.var}{TBW}

\item{spatial.pars}{TBW}

\item{method}{TBW}

\item{family.link}{TBW}

\item{global.nc.attributes}{TBW}
}
\value{
TBW
}
\description{
Downscale CMIP5 grid data
}
\examples{
#TBW
}
