% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{loadHistoricalTraceGrid}
\alias{loadHistoricalTraceGrid}
\title{Load TraCE21ka data file for a single variable.}
\usage{
loadHistoricalTraceGrid(
  file,
  var = NULL,
  lon_lim = c(-25, 25),
  lat_lim = c(25, 50),
  years = NULL,
  dictionary = system.file("extdata", "TraCE21ka_dictionary.csv", package = "dsclim")
)
}
\arguments{
\item{file}{String with the name of the file (including relative path if not in the working directory).}

\item{var}{String indicating the name of the variable that should be loaded.}

\item{lon_lim}{Vector with two integers, indicating the longitudinal extent of the desired study area.}

\item{lat_lim}{Vector with two integers, indicating the latitudinal extent of the desired study area.}

\item{years}{Integer vector indicating the years that should be loaded.}

\item{dictionary}{String indicating the path to a dictionary for the TraCE21ka dataset. Default is a predefined dictionary.}
}
\value{
List object with the structure defined in the loadeR package.
}
\description{
Load TraCE21ka data file for a single variable.
}
\examples{
\dontrun{
tas <- loadHistoricalTraceGrid(
  "Data/TraCE21ka/TS/trace.36.400BP-1990CE.cam2.h0.TS.2160101-2204012.nc",
  "tas"
)
}
}
