% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{loadHistoricalTraceGrid}
\alias{loadHistoricalTraceGrid}
\title{Title}
\usage{
loadHistoricalTraceGrid(
  file,
  var = NULL,
  lonLim = c(-25, 25),
  latLim = c(25, 50),
  start_date = "1551-01-01",
  end_date = "1990-12-31",
  years = 1961:1990,
  dictionary = system.file("extdata", "TraCE21ka_dictionary.csv", package =
    "TraCE21kaDSR")
)
}
\arguments{
\item{file}{TBW}

\item{var}{TBW}

\item{lonLim}{TBW}

\item{latLim}{TBW}

\item{start_date}{TBW}

\item{end_date}{TBW}

\item{years}{TBW}

\item{dictionary}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
# tas <- loadHistoricalTraceGrid(
#            "Data/TraCE21ka/TS/trace.36.400BP-1990CE.cam2.h0.TS.2160101-2204012.nc",
#            "tas")
}
