% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{loadTrace}
\alias{loadTrace}
\title{Title}
\usage{
loadTrace(
  files,
  vars = NULL,
  lon_lim = c(-25, 25),
  lat_lim = c(25, 50),
  years = NULL,
  compute_wss = TRUE,
  selection_vars = c("tas", "tasmax", "tasmin", "hurs@992.5561", "ps", "pr", "cld",
    "wss"),
  dictionary = system.file("extdata", "TraCE21ka_dictionary.csv", package = "dsclim")
)
}
\arguments{
\item{files}{TBW}

\item{vars}{TBW}

\item{lon_lim}{TBW}

\item{lat_lim}{TBW}

\item{years}{TBW}

\item{compute_wss}{TBW}

\item{selection_vars}{TBW}

\item{dictionary}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
# LGM.files <- traceFileNames(
#                         "Data/TraCE21ka/", 1)
# data <- loadTrace(LGM.files,
#                                  c("tas", "tasmax", "tasmin", "hurs@992.5561",
#                                  "ps", "pr", "cld", "u@992.5561", "v@992.5561"), 1,
#                                  selection_vars = c("tas", "tasmax", "tasmin",
#                                                  "hurs@992.5561", "ps", "pr",
#                                                  "cld", "wss"))
}
