% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{loadTraceGrid}
\alias{loadTraceGrid}
\title{Title}
\usage{
loadTraceGrid(
  file,
  var,
  lonLim = c(-25, 25),
  latLim = c(25, 50),
  dictionary = system.file("extdata", "TraCE21ka_dictionary.csv", package =
    "TraCE21kaDSR")
)
}
\arguments{
\item{file}{TBW}

\item{var}{TBW}

\item{lonLim}{TBW}

\item{latLim}{TBW}

\item{dictionary}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
# tas <- loadTraceGrid(
#            "Data/TraCE21ka/TS/trace.01.22000-20001BP.cam2.h0.TS.0000101-0200012.nc",
#            "tas", 1, lonLim = NULL, latLim=NULL) 
}
