% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UERRA_functions.R
\name{loadUerra}
\alias{loadUerra}
\title{Load UERRA data into climate4R grid format}
\usage{
loadUerra(
  file,
  var,
  lonLim = NULL,
  latLim = NULL,
  dictionary = system.file("extdata", "UERRA_dictionary.csv", package = "TraCE21kaDSR"),
  aggr.times = NULL,
  aggr.fun = mean
)
}
\arguments{
\item{file}{A string with the file name with the UERRA data to be loaded.}

\item{var}{A string with the variable name from the file to be loaded.}

\item{lonLim}{An optional vector with two values specifying the longitudinal limits (min and max) to be loaded.}

\item{latLim}{An optional vector with two values specifying the latitudinal limits (min and max) to be loaded.}

\item{dictionary}{An optional string with the filename with the dictionary for the UERRA data. By default it loads an internal dictionary from the package.}

\item{aggr.times}{An optional numeric specifying the number of cells to be aggregated.}

\item{aggr.fun}{An optional function specifying the function to be used in aggregation if aggr.times was specified.}
}
\value{
TBW
}
\description{
Load UERRA data into climate4R grid format
}
\examples{
# uerra.file <- "../Data/UERRA/UERRA-HARMONIE/2m_temperature/latlon/1961-90_2m_temperature.nc"
# local.var <- "tas"
# uerra = loadUerra(uerra.file, local.var)
}
