% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_climate4R_functions.R
\name{modifyDates}
\alias{modifyDates}
\title{Title}
\usage{
modifyDates(x, start_date = "1961-01-01", end_date = "1990-12-31")
}
\arguments{
\item{x}{TBW}

\item{start_date}{TBW}

\item{end_date}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
#TBW
}
