% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TraCE21ka_functions.R
\name{traceFileNames}
\alias{traceFileNames}
\title{Title}
\usage{
traceFileNames(indir, files_n = 36, vars = NULL)
}
\arguments{
\item{indir}{TBW}

\item{files_n}{TBW}

\item{vars}{TBW}
}
\value{
TBW
}
\description{
Title
}
\examples{
# historical.files <- traceFileNames("Data/TraCE21ka/", 36)
# LGM.files <- traceFileNames("Data/TraCE21ka/", 1)
}
